+++ jmx/src/main/java/uk/gov/justice/services/jmx/command/SystemCommanderMBean.java
@@ -1,9 +1,0 @@
-package uk.gov.justice.services.jmx.command;
-
-import javax.management.MXBean;
-
-@MXBean
-public interface SystemCommanderMBean {
-
-    void runCommand(final SystemCommand systemCommand);
-}